HuTu 80/requirements.txt
pandas
numpy
matplotlib
seaborn
Pillow
opencv-python
scipy
scikit-learn
torch
torchvision
segmentation-models-pytorch
torchinfo
tqdm